+++ app/controllers/api/v1/projects_controller.rb
@@ -7,4 +7,10 @@ class Api::V1::ProjectsController < Api::V1::BaseController
   # # GET /api/v1/projects/1-AUTHENTICATION_TOKEN.json
   # def show
   # end
+
+  # GET /api/v1/projects/1-AUTHENTICATION_TOKEN/survey-info.json
+  def survey_info
+    @event = @project.events.find_by_param(params[:event])
+    @design = @project.designs.find_by_param(params[:design])
+  end
 end