+++ test/controllers/api/v1/projects_controller_test.rb
@@ -13,4 +13,11 @@ class Api::V1::ProjectsControllerTest < ActionDispatch::IntegrationTest
     assert_not_nil assigns(:project)
     assert_response :success
   end
+
+  test "should get survey info" do
+    get api_v1_survey_info_path(authentication_token: @project.id_and_token, event: events(:api_event), design: designs(:api_design), format: "json")
+    assert_not_nil assigns(:event)
+    assert_not_nil assigns(:design)
+    assert_response :success
+  end
 end